+++ src/main/java/org/ggp/base/player/gamer/statemachine/sample/SampleMonteCarloGamer.java
@@ -52,7 +52,7 @@ public final class SampleMonteCarloGamer extends SampleGamer
     		    if (System.currentTimeMillis() > finishBy)
     		        break;
 
-    		    int theScore = performDepthChargeFromMove(getCurrentState(), moves.get(i));
+    		    double theScore = performDepthChargeFromMove(getCurrentState(), moves.get(i));
     		    moveTotalPoints[i] += theScore;
     		    moveTotalAttempts[i] += 1;
     		    visitedNodes += this.depth[0] + 1;
@@ -89,7 +89,7 @@ public final class SampleMonteCarloGamer extends SampleGamer
 	}
 
 	private int[] depth = new int[1];
-	int performDepthChargeFromMove(ExplicitMachineState theState, ExplicitMove myMove) {
+	double performDepthChargeFromMove(ExplicitMachineState theState, ExplicitMove myMove) {
 	    StateMachine theMachine = getStateMachine();
 	    try {
             ExplicitMachineState finalState = theMachine.performDepthCharge(theMachine.getRandomNextState(theState, getRole(), myMove), depth);
